AI-GF/requirements.txt
aiofiles
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
attrs
av
certifi
cffi
charset-normalizer
click
colorama
coloredlogs
distro
docstring_parser
eval_type_backport
filelock
flatbuffers
frozenlist
fsspec
fastapi
uvicorn[standard]
h11
httpcore
httpx
huggingface-hub
humanfriendly
idna
Jinja2
jiter
livekit
livekit-agents
livekit-api
livekit-plugins-cartesia
livekit-plugins-elevenlabs
livekit-plugins-noise-cancellation
livekit-plugins-anthropic
livekit-plugins-google
livekit-plugins-openai
livekit-plugins-rime
livekit-plugins-silero
livekit-plugins-turn-detector
livekit-protocol
MarkupSafe
mpmath
multidict
nest-asyncio
numpy
onnxruntime
openai
packaging
pillow
propcache
protobuf
psutil
pycparser
pydantic
pydantic_core
PyJWT
python-dotenv
PyYAML
regex
requests
safetensors
sniffio
sounddevice
sympy
tokenizers
torch
tqdm
transformers
types-protobuf
typing-inspection
typing_extensions
Unidecode
urllib3
watchfiles
websockets
yarl
